@@ -19,8 +19,8 @@
     "google-protobuf": "^3.21.0",
     "node-gettext": "^3.0.0",
     "rbush": "^2.0.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^7.2.2",
     "react-router": "^5.1.2",
     "react-simple-maps": "^3.0.0",
@@ -45,8 +45,8 @@
     "@types/node": "^16.11.26",
     "@types/node-gettext": "^3.0.3",
     "@types/rbush": "^2.0.2",
-    "@types/react": "^16.9.56",
-    "@types/react-dom": "^16.9.9",
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6",
     "@types/react-redux": "^7.1.24",
     "@types/react-router": "^5.1.19",
     "@types/react-simple-maps": "^1.0.8",
